Split utilities whose boundaries depended on one another
