+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/error/programErrorVoictent.ts
@@ -5,7 +5,7 @@ import {
   ProgramFileCache,
   SerializedHubblepup,
 } from '../../../layer-agnostic-utilities/program/programFileCache';
-import { serialize } from '../../../package-agnostic-utilities/typed-datum/serializer/serialize';
+import { serialize } from '../../../package-agnostic-utilities/one-way-serializer/serialize';
 import { FileExtensionSuffixIdentifier } from '../../../package-agnostic-utilities/file/fileExtensionSuffixIdentifier';
 import { AbstractAsymmetricInMemoryVoictent2 } from '../in-memory-cache/abstractInMemoryVoictent2';
 import {
