+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/linting/lintAssertion.ts
@@ -3,7 +3,7 @@ import {
   GenericZorn2Template,
   Zorn2,
 } from '../../../package-agnostic-utilities/datastructure/zorn';
-import { TypeScriptObject } from '../../../package-agnostic-utilities/typed-datum/type-script/object';
+import { TypeScriptObject } from '../../../package-agnostic-utilities/object/object';
 import {
   MessageAccessorInputFromAccessor,
   RuleZorn,
