+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/function/function.ts
@@ -1,3 +1,3 @@
-import { Tuple } from '../../type/tuple';
+import { Tuple } from '../type/tuple';
 
 export type TypeScriptFunction = (...argumentTuple: Tuple<unknown>) => unknown;
