+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/object/getPrototypeNameTuple.ts
