+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/one-way-serializer/getSerializeableNodeLineList.ts
@@ -1,4 +1,4 @@
-import { CustomDatumTypeName } from '../customTypedDatum';
+import { CustomDatumTypeName } from '../typed-datum/customTypedDatum';
 import {
   SerializeableNode,
   SerializeableNodeName,
