+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/one-way-serializer/serialize.ts
