+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/typed-datum/typeScriptTypedDatum.ts
@@ -1,5 +1,5 @@
-import { TypeScriptFunction } from './function';
-import { TypeScriptObject } from './object';
+import { TypeScriptFunction } from '../function/function';
+import { TypeScriptObject } from '../object/object';
 
 export enum TypeScriptDatumTypeName {
   bigint = 'bigint',
